feat: usize retirement stages 3+4 -- the type is GONE
Stage 3: wiki/chapters swept (13 mechanical files usize -> u64; the
index/count/address brief -- the settled semantic source -- gets migration
status instead; the appendix keeps its retirement narration).

Stage 4, the compiler rejection:
- PrimitiveType::Usize/Isize DELETED from both tree representations, with
  every match arm swept; "usize"/"isize"/"AtomicUsize" removed from the
  builtin symbol table (AtomicU64 already existed; the one atomics canary
  updated). BUILTIN_TYPE_COUNT 28 -> 25.
- BuiltinType::ordinal() hardcodes symbol-table POSITIONS and silently
  broke on the removal ("unknown layout-bearing type UInt" -- the dungeon
  family caught it); ordinals corrected and the coupling PINNED by a unit
  test that walks the real table (which caught my own first guess too).
- migration tolerances collapsed: wire contracts are u64-only, the
  termination natural-class comparison is direct (len/subtract synthesize
  "u64" now), the proof-fact "usize" arm is gone.
- the std/core/host/lattice-corpus .omg trees swept too (the recipe's
  inventory had missed them -- the FilesystemHost boundary traits and core
  containers all carried usize). region.omg's allocate/deallocate address
  slots are annotated as the addr follow-up (trait + providers + callers
  move together).
- fail canaries types/usize_rejected + types/isize_rejected pin the loud
  unknown-type diagnostic.
- unit-fixture fallout repaired across checked-trees + validation (Omega
  source blocks swept inside r#""# only -- a blind sweep hits Rust's own
  usize). Two tests found ALREADY ROTTED on clean origin/main (this
  crate's tests sit outside the root gate): the ambiguous-operator
  message assertions matched a long-gone numeric format, and the
  dynamic-indexed requires fixture lacked its index field entirely; both
  repaired -- 164/164 green.

Gate: canary_suite 776/0, differential 13/13, core 60/60, all touched
crates' unit tests green. Pre-existing parallel-lane reds unchanged
(stdin trio; led_mixer .gitignore structure tests).

+++ compiler/omega-rs/pipeline/omega-typed-trees-to-checked-trees/src/tests/borrow/checks.rs
@@ -677,7 +677,7 @@ fn rejects_mutating_call_through_owner_while_view_is_active() {
         }
 
         machine Main::read_alias(&self, entries: &[Entry]) {
-            let count: usize = entries.len;
+            let count: u64 = entries.len;
         }
     "#;
 
@@ -808,7 +808,7 @@ fn accepts_mutating_call_through_owner_on_disjoint_field() {
         }
 
         machine Main::read_alias(&self, entries: &[Entry]) {
-            let count: usize = entries.len;
+            let count: u64 = entries.len;
         }
     "#;
 
@@ -868,7 +868,7 @@ fn accepts_known_pure_mutable_receiver_call_while_view_is_active() {
         }
 
         machine Main::read_alias(&self, entries: &[Entry], value: i32) {
-            let count: usize = entries.len;
+            let count: u64 = entries.len;
         }
     "#;
 
@@ -966,9 +966,9 @@ fn accepts_recursive_slice_parameter_index_proof_from_guard() {
                 _ -> self.visit(entries, 0)
             }
 
-            state visit(&mut self, entries: &[Entry], index: usize) {
+            state visit(&mut self, entries: &[Entry], index: u64) {
                 let value: i32 = entries[index].value;
-                let next_index: usize = index + 1;
+                let next_index: u64 = index + 1;
                 let has_next: bool = next_index < entries.len;
 
                 transition {
